Direct benchmarks for matching, trade_executor and order_processing (#2138)
