+++ lib/peatio/bench/helpers.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Bench
+  module Helpers
+    def become_billionaire(member)
+      @currencies.each do |c|
+        Factories.create(:deposit, member_id: member.id, currency_id: c.id)
+      end
+    end
+  end
+end
