+++ lib/peatio/bench/injectors.rb
