+++ lib/peatio/bench/injectors/bitfinex.rb
