+++ lib/tasks/bench.rake
@@ -4,66 +4,139 @@
 # TODO: Don't sleep in case of last bench.
 # TODO: Remove legacy benchmarks.
 namespace :bench do
-  desc 'Matching'
-  task :matching, [:config_load_path] => [:environment] do |_t, args|
-    args.with_defaults(:config_load_path => 'config/bench/matching.yml')
+  
+  namespace :matching do
 
-    benches =
-      YAML.load_file(Rails.root.join(args[:config_load_path]))
-        .map(&:deep_symbolize_keys)
-        .each_with_object([]) do |config, memo|
+  desc 'Matching with amqp messages'  
+    task :amqp, [:config_load_path] => [:environment] do |_t, args|
+      args.with_defaults(:config_load_path => 'config/bench/matching.yml')
+
+      benches =
+        YAML.load_file(Rails.root.join(args[:config_load_path]))
+          .map(&:deep_symbolize_keys)
+          .each_with_object([]) do |config, memo|
+            Kernel.pp config
+
+            matching = Bench::Matching::AMQP.new(config)
+            matching.run!
+            memo << matching
+            matching.save_report
+            Kernel.puts "Sleep before next bench"
+            sleep 5
+          end
+
+      benches.each {|b| Kernel.pp b.result}
+    end
+
+    desc 'Matching without amqp messages'
+    task :direct, [:config_load_path] => [:environment] do |_t, args|
+      args.with_defaults(:config_load_path => 'config/bench/matching.yml')
+
+      benches =
+        YAML.load_file(Rails.root.join(args[:config_load_path]))
+          .map(&:deep_symbolize_keys)
+          .each_with_object([]) do |config, memo|
+            Kernel.pp config
+
+            matching = Bench::Matching::Direct.new(config)
+            matching.run!
+            memo << matching
+            matching.save_report
+            Kernel.puts "Sleep before next bench"
+            sleep 5
+          end
+
+      benches.each {|b| Kernel.pp b.result}
+    end
+  end
+
+  namespace :trade_execution do
+
+    desc 'Trade Execution with amqp messages'
+    task :amqp, [:config_load_path] => [:environment] do |_t, args|
+      args.with_defaults(:config_load_path => 'config/bench/trade_execution.yml')
+
+      benches =
+        YAML.load_file(Rails.root.join(args[:config_load_path]))
+          .map(&:deep_symbolize_keys)
+          .each_with_object([]) do |config, memo|
           Kernel.pp config
 
-          matching = Bench::Matching.new(config)
-          matching.run!
-          memo << matching
-          matching.save_report
+          trade_execution = Bench::TradeExecution::AMQP.new(config)
+          trade_execution.run!
+          memo << trade_execution
+          trade_execution.save_report
           Kernel.puts "Sleep before next bench"
           sleep 5
         end
 
-    benches.each {|b| Kernel.pp b.result}
+      benches.each {|b| Kernel.pp b.result}
+    end
+
+    desc 'Trade execution without amqp messages'
+    task :direct, [:config_load_path] => [:environment] do |_t, args|
+      args.with_defaults(:config_load_path => 'config/bench/trade_execution.yml')
+
+      benches =
+        YAML.load_file(Rails.root.join(args[:config_load_path]))
+          .map(&:deep_symbolize_keys)
+          .each_with_object([]) do |config, memo|
+          Kernel.pp config
+
+          trade_execution = Bench::TradeExecution::Direct.new(config)
+          trade_execution.run!
+          memo << trade_execution
+          trade_execution.save_report
+          Kernel.puts "Sleep before next bench"
+          sleep 5
+        end
+
+      benches.each {|b| Kernel.pp b.result}
+    end
   end
 
-  desc 'Trade Execution'
-  task :trade_execution, [:config_load_path] => [:environment] do |_t, args|
-    args.with_defaults(:config_load_path => 'config/bench/trade_execution.yml')
+  namespace :order_processing do
 
-    benches =
-      YAML.load_file(Rails.root.join(args[:config_load_path]))
-        .map(&:deep_symbolize_keys)
-        .each_with_object([]) do |config, memo|
-        Kernel.pp config
+    desc 'Order Processing with amqp messages'
+    task :amqp, [:config_load_path] => [:environment] do |_t, args|
+      args.with_defaults(:config_load_path => 'config/bench/order_processing.yml')
 
-        trade_execution = Bench::TradeExecution.new(config)
-        trade_execution.run!
-        memo << trade_execution
-        trade_execution.save_report
-        Kernel.puts "Sleep before next bench"
-        sleep 5
-      end
+      benches =
+        YAML.load_file(Rails.root.join(args[:config_load_path]))
+          .map(&:deep_symbolize_keys)
+          .each_with_object([]) do |config, memo|
+          Kernel.pp config
 
-    benches.each {|b| Kernel.pp b.result}
-  end
+          order_processing = Bench::OrderProcessing::AMQP.new(config)
+          order_processing.run!
+          memo << order_processing
+          order_processing.save_report
+          Kernel.puts "Sleep before next bench"
+          sleep 5
+        end
 
-  desc 'Order Processing'
-  task :order_processing, [:config_load_path] => [:environment] do |_t, args|
-    args.with_defaults(:config_load_path => 'config/bench/order_processing.yml')
+      benches.each {|b| Kernel.pp b.result}
+    end
 
-    benches =
-      YAML.load_file(Rails.root.join(args[:config_load_path]))
-        .map(&:deep_symbolize_keys)
-        .each_with_object([]) do |config, memo|
-        Kernel.pp config
+    desc 'Order Processing without amqp messages'
+    task :direct, [:config_load_path] => [:environment] do |_t, args|
+      args.with_defaults(:config_load_path => 'config/bench/order_processing.yml')
 
-        order_processing = Bench::OrderProcessing.new(config)
-        order_processing.run!
-        memo << order_processing
-        order_processing.save_report
-        Kernel.puts "Sleep before next bench"
-        sleep 5
-      end
+      benches =
+        YAML.load_file(Rails.root.join(args[:config_load_path]))
+          .map(&:deep_symbolize_keys)
+          .each_with_object([]) do |config, memo|
+          Kernel.pp config
 
-    benches.each {|b| Kernel.pp b.result}
+          order_processing = Bench::OrderProcessing::Direct.new(config)
+          order_processing.run!
+          memo << order_processing
+          order_processing.save_report
+          Kernel.puts "Sleep before next bench"
+          sleep 5
+        end
+
+      benches.each {|b| Kernel.pp b.result}
+    end
   end
 end
